dart_detection_system/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
tk>=0.1.0
